+++ zzre/tools/sceneeditor/SceneEditor.FOModel.cs
@@ -54,7 +54,7 @@ public partial class SceneEditor
                 (material.Texture.Texture, material.Sampler.Sampler) = textureLoader.LoadTexture(textureBasePaths, rwMaterial);
                 material.LinkTransformsTo(camera);
                 material.World.BufferRange = locationRange;
-                material.Colors.Ref = ModelStandardMaterialUniforms.Default;
+                material.Colors.Ref = ModelColors.Default;
                 material.Colors.Ref.vertexColorFactor = 0.0f;
                 material.Colors.Ref.tint = rwMaterial.color.ToFColor() * sceneModel.color;
                 return material;
